new/multi_bot/requirements.txt
langchain==0.1.4
langchain_community==0.0.13
langchain_google_genai==0.0.4
google-generativeai>=0.3.0
google-cloud-speech>=2.25.1
sentence-transformers>=2.2.2
chromadb==0.4.18
pypdf==3.17.0
streamlit==1.28.0
python-dotenv==1.0.0
fastapi>=0.101.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
httpx>=0.24.0
flask==2.3.3
twilio==8.5.0
requests==2.31.0
